back_end/requirements.txt
asgiref==3.4.1
Django==3.2.7
django-cors-headers==3.8.0
django-redis-cache==3.0.0
djangorestframework==3.12.4
Pillow==8.3.2
psycopg2-binary==2.9.1
pytz==2021.1
redis==3.5.3
sqlparse==0.4.2
Django>=3.0,<4.0
psycopg2-binary>=2.8